+++ examples/datadog_json_logger.rs
@@ -9,7 +9,7 @@ async fn main() {
             .with_env(Environment::Dev)
             .with_version("1.0".to_string())
             // We need a tracer if we want trace and span IDs to be created and propagated, otherwise logs won't contain these correlation IDs
-            // You can also setup custom tracer and custom subscriber if you don't wanna use the `tracing` feature
+            // You can also setup custom tracer and custom subscriber if you don't wanna use the `traces` feature
             .with_telemetry("http://localhost:55681/v1/traces".to_string(), service_name)
             .build(),
     );